Flexnet improvements - Add devnet image integration to jenkins CI - Start rootless docker if not running - Remove config files from codebase flexnet0 - Don't overwrite scripts in script-gen - Reduce configs/scripts in codebase flexnet1 - Update blst submodule if not initialized

+++ docker/flexnet/common/ethtx-gen.py
@@ -44,5 +44,7 @@ for region in regions:
         cwd = Path(args.topology_input_file).parents[0]
         ethtx_path = cwd / volume / "scripts" / "ethtx.sh"
         ethtx_path.parent.mkdir(parents=True, exist_ok=True)
+        if ethtx_path.exists():
+            continue
         with open(ethtx_path, "w+") as ethtx_file:
             ethtx_file.write(ethtx_text)
